move ManagedFile to lib
