+++ src/package.rs
@@ -36,9 +36,9 @@ use crate::{
     credentials::Credentials,
     errors::{DeserializationError, SerializationError},
     lock::{LockedPackage, Lockfile},
-    managed_file::ManagedFile,
     manifest::{self, Dependency, Manifest, MANIFEST_FILE},
     registry::{Artifactory, RegistryUri},
+    ManagedFile,
 };
 
 /// IO abstraction layer over local `buffrs` package store
